Renamed pausemenu header
